+++ boards/wio_terminal/src/sensors.rs
@@ -2,7 +2,10 @@ use atsamd_hal::adc::Adc;
 use atsamd_hal::clock::GenericClockController;
 use atsamd_hal::pac::gclk::pchctrl::GEN_A::GCLK11;
 use atsamd_hal::pac::{ADC1, MCLK, SERCOM4};
-use atsamd_hal::sercom::{I2CMaster4, PadPin, Sercom4Pad0, Sercom4Pad1};
+use atsamd_hal::sercom::{
+    v2::{i2c, IoSet3, Sercom4},
+    PadPin,
+};
 use atsamd_hal::time::U32Ext;
 
 use lis3dh::{Lis3dh, SlaveAddr};
@@ -18,6 +21,11 @@ pub struct Accelerometer {
     pub sda: I2c0SdaReset,
 }
 
+/// I2C pads for the labelled I2C peripheral
+///
+/// You can use these pads with other, user-defined [`i2c::Config`]urations.
+pub type I2cPads = i2c::Pads<Sercom4, IoSet3, I2c0Sda, I2c0Scl>;
+
 impl Accelerometer {
     /// Initialize the LIS3DH accelerometer using the correct pins and
     // peripherals. Use the driver's default settings.
@@ -26,19 +34,18 @@ impl Accelerometer {
         clocks: &mut GenericClockController,
         sercom4: SERCOM4,
         mclk: &mut MCLK,
-    ) -> Lis3dh<I2CMaster4<I2c0Sda, I2c0Scl>> {
+    ) -> Lis3dh<i2c::I2c<i2c::Config<I2cPads>>> {
         // The accelerometer is connected to the Wio Terminal's `I2C0` bus, so
         // based on the possible padouts listed in the datasheet it must use
-        // `SERCOM4` and in turn `I2CMaster4`.
+        // `SERCOM4`.
         let gclk0 = clocks.gclk0();
-        let i2c = I2CMaster4::new(
-            &clocks.sercom4_core(&gclk0).unwrap(),
-            400.khz(),
-            sercom4,
-            mclk,
-            self.sda.into(),
-            self.scl.into(),
-        );
+        let clock = &clocks.sercom4_core(&gclk0).unwrap();
+        let freq = clock.freq();
+        let (sda, scl): (I2c0Sda, I2c0Scl) = (self.sda.into(), self.scl.into());
+        let pads: I2cPads = i2c::Pads::new(sda, scl);
+        let i2c = i2c::Config::new(mclk, sercom4, pads, freq)
+            .baud(400.khz())
+            .enable();
 
         // The schematic states that the alternate I2C address `0x19` is used,
         // but that doesn't appear to work!