+++ hal/src/dmac/transfer.rs
@@ -512,6 +512,7 @@ where
         }
     }
 }
+
 impl<B, C, const N: usize> Transfer<C, BufferPair<&'static mut [B; N]>>
 where
     B: 'static + Beat,
@@ -546,6 +547,29 @@ where
     #[inline]
     pub fn software_trigger(&mut self) {
         self.chan.as_mut().software_trigger();
+    }
+
+    /// Unsafely and mutably borrow the source buffer
+    ///
+    /// # Safety
+    ///
+    /// The source buffer should never be borrowed when a transfer is in
+    /// progress, as it is getting mutated or read in another context (ie,
+    /// the DMAC hardware "thread").
+    #[inline]
+    pub(crate) unsafe fn borrow_source(&mut self) -> &mut S {
+        &mut self.buffers.source
+    }
+
+    /// Unsafely and mutably borrow the destination buffer.
+    /// # Safety
+    ///
+    /// The destination buffer should never be borrowed when a transfer is in
+    /// progress, as it is getting mutated or read in another context (ie,
+    /// the DMAC hardware "thread").
+    #[inline]
+    pub(crate) unsafe fn borrow_destination(&mut self) -> &mut D {
+        &mut self.buffers.destination
     }
 
     /// Wait for the DMA transfer to complete and release all owned