+++ boards/metro_m0/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- Update `i2c_master` convenience function to use the new `sercom::v2::i2c` API
+- Add an `i2c` example
 - Updated to 2021 edition, updated dependencies, removed unused dependencies (#562)
 
 # v0.11.0