+++ boards/metro_m4/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- Update `i2c_master` convenience function to use the new `sercom::v2::i2c` API
+- Add an `i2c` example
 - Fix incorrect clocking in `uart` helper function
 - Updated to 2021 edition, updated dependencies, removed unused dependencies (#562)